Graphics classification optimized
The previous version of the classifier was inneficient with large texts, so this part has been partially reimplemented.
By now, the custom_text module has been reimplemented and renamed as main_text, the separator module has been modified, and a new abstraction layer (separator_handler) has been added to manage all the behaviour between the separators.
Also, the rects and descriptor modules has been partially reimplemented adding a new abstraction layer. With the new approach of this module, the troughput of the app is the same no matter the text lenght.